+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - support for `ruff` style checker
 - `get_source_code()` function in audit.py
 - `count_comments()` function in audit.py
+- `within_deadline()` function in common.py
+- `postpone_tests()` function in common.py
 
 ### Fixed